refactor(svelte): extract 5 components and remove 4 dead component files
Extract BottomBar (with git panel), SettingsModal, UsageDashboard,
DeleteConfirmModal, and UpdateToast from +page.svelte into focused
component files. Remove unused Sidebar, Terminal, NewSessionModal,
and ContextMenu components.
