+++ src/lib/components/DeleteConfirmModal.svelte
@@ -1,0 +1,35 @@
+<script>
+  let {
+    profile,
+    onConfirm,
+    onCancel,
+  } = $props();
+</script>
+
+{#if profile}
+<div class="modal-backdrop" style="animation:fadeIn 0.1s ease-out;">
+  <div class="modal" style="max-width:360px;animation:slideIn 0.15s ease-out;">
+    <div style="text-align:center;padding:20px 20px 0;">
+      <svg width="32" height="32" viewBox="0 0 16 16" fill="#f85149" style="margin-bottom:12px;"><path d="M6.5 1.75a.25.25 0 01.25-.25h2.5a.25.25 0 01.25.25V3h-3V1.75zM11 3V1.75A1.75 1.75 0 009.25 0h-2.5A1.75 1.75 0 005 1.75V3H2.75a.75.75 0 000 1.5h.928l.747 10.218A1.75 1.75 0 006.172 16h3.656a1.75 1.75 0 001.747-1.282L12.322 4.5h.928a.75.75 0 000-1.5H11zm-5.522 1.5l.735 10.06a.25.25 0 00.249.19h3.076a.25.25 0 00.249-.19l.735-10.06H5.478z"/></svg>
+      <h2 style="font-size:15px;margin-bottom:8px;">Delete this session?</h2>
+      <p style="font-size:13px;color:var(--text-secondary);line-height:1.5;">
+        Are you sure you want to delete the <strong style="color:var(--text-primary);">{profile.projectName} — {profile.purpose}</strong> session?
+      </p>
+    </div>
+    <div class="modal-actions" style="padding:16px 20px;">
+      <button onclick={onCancel}>Cancel</button>
+      <button style="background:#f85149 !important;border-color:transparent !important;color:#fff !important;" onclick={onConfirm}>Delete</button>
+    </div>
+  </div>
+</div>
+{/if}
+
+<style>
+  .modal-backdrop { position: fixed; inset: 0; background: rgba(0,0,0,0.6); display: flex; align-items: center; justify-content: center; z-index: 1000; animation: fadeIn 0.15s ease-out; }
+  .modal { background: var(--modal-bg, #161b22); border: 1px solid var(--border); border-radius: 12px; padding: 20px; width: 420px; max-width: 90vw; animation: slideIn 0.2s ease-out; }
+  @keyframes fadeIn { from { opacity: 0; } to { opacity: 1; } }
+  @keyframes slideIn { from { opacity: 0; transform: translateY(-10px) scale(0.98); } to { opacity: 1; transform: translateY(0) scale(1); } }
+  .modal h2 { font-size: 15px; color: var(--text-primary); margin: 0 0 16px; }
+  .modal-actions { display: flex; justify-content: flex-end; gap: 8px; margin-top: 16px; }
+  .modal-actions button { padding: 7px 16px; border-radius: 6px; font-size: 13px; cursor: pointer; border: 1px solid var(--border); background: var(--btn-bg, #21262d); color: var(--text-primary); font-family: inherit; }
+</style>
