+++ src/routes/+page.svelte
@@ -13,6 +13,11 @@
   import { gitStore } from "$lib/stores/git.svelte";
   import { terminalStore } from "$lib/stores/terminal.svelte";
   import { shellStore } from "$lib/stores/shell.svelte";
+  import DeleteConfirmModal from "$lib/components/DeleteConfirmModal.svelte";
+  import UpdateToast from "$lib/components/UpdateToast.svelte";
+  import BottomBar from "$lib/components/BottomBar.svelte";
+  import SettingsModal from "$lib/components/SettingsModal.svelte";
+  import UsageDashboard from "$lib/components/UsageDashboard.svelte";
 
   let profiles = $state([]);
   let activeProfile = $state(null);
@@ -1004,219 +1009,9 @@ Anti-patterns to avoid:
     </div>
   </div>
 </div>
-<div class="bottom-bar">
-  <div class="bottom-left">
-    <div class="profile-wrap">
-      <!-- svelte-ignore a11y_no_static_element_interactions a11y_click_events_have_key_events -->
-      <div onclick={(e) => { e.stopPropagation(); profileMenuOpen = !profileMenuOpen; }}>
-        <button class="profile-avatar" title="Profile"><span class="avatar-letter">CG</span></button>
-      </div>
-      {#if profileMenuOpen}
-        <!-- svelte-ignore a11y_no_static_element_interactions a11y_click_events_have_key_events -->
-        <div class="profile-menu" onclick={(e) => e.stopPropagation()}>
-          <button class="pm-item" onclick={() => { profileMenuOpen = false; showSettings = true; settingsTab = 'settings'; }}>
-            <svg viewBox="0 0 24 24"><circle cx="12" cy="12" r="3"/><path d="M12.22 2h-.44a2 2 0 00-2 2v.18a2 2 0 01-1 1.73l-.43.25a2 2 0 01-2 0l-.15-.08a2 2 0 00-2.73.73l-.22.38a2 2 0 00.73 2.73l.15.1a2 2 0 011 1.72v.51a2 2 0 01-1 1.74l-.15.09a2 2 0 00-.73 2.73l.22.38a2 2 0 002.73.73l.15-.08a2 2 0 012 0l.43.25a2 2 0 011 1.73V20a2 2 0 002 2h.44a2 2 0 002-2v-.18a2 2 0 011-1.73l.43-.25a2 2 0 012 0l.15.08a2 2 0 002.73-.73l.22-.39a2 2 0 00-.73-2.73l-.15-.08a2 2 0 01-1-1.74v-.5a2 2 0 011-1.74l.15-.09a2 2 0 00.73-2.73l-.22-.38a2 2 0 00-2.73-.73l-.15.08a2 2 0 01-2 0l-.43-.25a2 2 0 01-1-1.73V4a2 2 0 00-2-2z"/></svg>
-            Settings
-          </button>
-          <button class="pm-item" onclick={() => { profileMenuOpen = false; showSettings = true; settingsTab = 'plugins'; pluginsStore.loadClaudePlugins(); }}>
-            <svg viewBox="0 0 24 24"><path d="M12 2L2 7l10 5 10-5-10-5z"/><path d="M2 17l10 5 10-5"/><path d="M2 12l10 5 10-5"/></svg>
-            Plugins
-          </button>
-          <button class="pm-item" onclick={() => { profileMenuOpen = false; usageStore.showDashboard = true; usageStore.loadDashboard(); }}>
-            <svg viewBox="0 0 24 24"><path d="M18 20V10M12 20V4M6 20v-6"/></svg>
-            Usage Dashboard
-          </button>
-          <div class="pm-sep"></div>
-          <button class="pm-item" onclick={() => { profileMenuOpen = false; openExternal('https://clauge.ssh-i.in/changelog.html'); }}>
-            <svg viewBox="0 0 24 24"><path d="M12 2L15.09 8.26 22 9.27 17 14.14 18.18 21.02 12 17.77 5.82 21.02 7 14.14 2 9.27 8.91 8.26z"/></svg>
-            What's New
-            <svg class="pm-external" viewBox="0 0 24 24"><path d="M18 13v6a2 2 0 01-2 2H5a2 2 0 01-2-2V8a2 2 0 012-2h6"/><polyline points="15 3 21 3 21 9"/><line x1="10" y1="14" x2="21" y2="3"/></svg>
-          </button>
-          <button class="pm-item" onclick={() => { profileMenuOpen = false; openExternal('https://github.com/ansxuman/Clauge/issues'); }}>
-            <svg viewBox="0 0 24 24"><circle cx="12" cy="12" r="10"/><path d="M12 8v4M12 16h.01"/></svg>
-            Report Issue
-            <svg class="pm-external" viewBox="0 0 24 24"><path d="M18 13v6a2 2 0 01-2 2H5a2 2 0 01-2-2V8a2 2 0 012-2h6"/><polyline points="15 3 21 3 21 9"/><line x1="10" y1="14" x2="21" y2="3"/></svg>
-          </button>
-          <button class="pm-item pm-coffee" onclick={() => { profileMenuOpen = false; openExternal('https://buymeacoffee.com/ansxuman'); }}>
-            <svg viewBox="0 0 24 24"><path d="M17 8h1a4 4 0 110 8h-1"/><path d="M3 8h14v9a4 4 0 01-4 4H7a4 4 0 01-4-4V8z"/><line x1="6" y1="2" x2="6" y2="4"/><line x1="10" y1="2" x2="10" y2="4"/><line x1="14" y1="2" x2="14" y2="4"/></svg>
-            Buy Me a Coffee
-            <svg class="pm-external" viewBox="0 0 24 24"><path d="M18 13v6a2 2 0 01-2 2H5a2 2 0 01-2-2V8a2 2 0 012-2h6"/><polyline points="15 3 21 3 21 9"/><line x1="10" y1="14" x2="21" y2="3"/></svg>
-          </button>
-        </div>
-      {/if}
-    </div>
-    {#if activeProfile && gitStore.gitBranch}
-    <div class="git-status-wrap">
-      <!-- svelte-ignore a11y_no_static_element_interactions a11y_click_events_have_key_events -->
-      <div class="git-status-bar" onclick={(e) => { e.stopPropagation(); gitStore.gitPanelOpen = !gitStore.gitPanelOpen; }}>
-        <svg width="11" height="11" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2.5" stroke-linecap="round" stroke-linejoin="round"><line x1="6" y1="3" x2="6" y2="15"/><circle cx="18" cy="6" r="3"/><circle cx="6" cy="18" r="3"/><path d="M18 9a9 9 0 01-9 9"/></svg>
-        <span class="git-bar-branch">{gitStore.gitBranch}</span>
-        {#if gitStore.gitAhead > 0}
-          <span class="git-bar-ahead" title="{gitStore.gitAhead} unpushed">↑{gitStore.gitAhead}</span>
-        {/if}
-        {#if gitStore.gitBehind > 0}
-          <span class="git-bar-behind" title="{gitStore.gitBehind} to pull">↓{gitStore.gitBehind}</span>
-        {/if}
-        {#if gitStore.gitFiles.length > 0}
-          <span class="git-bar-changes">{gitStore.gitFiles.length}</span>
-        {/if}
-        {#if gitStore.gitMsg}
-          <span class="git-bar-msg">{gitStore.gitMsg}</span>
-        {/if}
-      </div>
-      {#if gitStore.gitPanelOpen}
-        <!-- svelte-ignore a11y_no_static_element_interactions a11y_click_events_have_key_events -->
-        <div class="git-popup" onclick={(e) => e.stopPropagation()}>
-          <div class="git-popup-tabs">
-            <button class="git-popup-tab" class:active={gitStore.gitTab === 'changes'} onclick={() => gitStore.gitTab = 'changes'}>Changes{gitStore.gitFiles.length > 0 ? ` (${gitStore.gitFiles.length})` : ''}</button>
-            <button class="git-popup-tab" class:active={gitStore.gitTab === 'history'} onclick={() => { gitStore.gitTab = 'history'; loadGitHistory(); }}>History</button>
-            <button class="git-popup-tab" class:active={gitStore.gitTab === 'branches'} onclick={() => { gitStore.gitTab = 'branches'; loadGitBranches(); }}>Branches</button>
-            <div class="git-popup-tab-actions">
-              <button class="git-action-btn has-tooltip" onclick={doGitStash}>
-                <svg width="12" height="12" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"><path d="M19 21H5a2 2 0 01-2-2V5a2 2 0 012-2h14a2 2 0 012 2v14a2 2 0 01-2 2z"/><path d="M17 21v-8H7v8"/><path d="M7 3v5h8"/></svg>
-                <span class="btn-tooltip">Stash</span>
-              </button>
-              <button class="git-action-btn has-tooltip" onclick={doGitStashPop}>
-                <svg width="12" height="12" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"><path d="M19 21H5a2 2 0 01-2-2V5a2 2 0 012-2h14a2 2 0 012 2v14a2 2 0 01-2 2z"/><polyline points="12 8 12 16"/><polyline points="8 12 12 8 16 12"/></svg>
-                <span class="btn-tooltip">Pop Stash</span>
-              </button>
-              <button class="git-action-btn has-tooltip" disabled={gitStore.gitLoading === 'pull'} onclick={doGitPull}>
-                <svg width="12" height="12" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"><line x1="12" y1="5" x2="12" y2="19"/><polyline points="19 12 12 19 5 12"/></svg>
-                <span class="btn-tooltip">{gitStore.gitLoading === 'pull' ? 'Pulling...' : 'Pull'}</span>
-              </button>
-              <button class="git-action-btn has-tooltip" disabled={gitStore.gitLoading === 'push'} onclick={doGitPush}>
-                <svg width="12" height="12" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"><line x1="12" y1="19" x2="12" y2="5"/><polyline points="5 12 12 5 19 12"/></svg>
-                <span class="btn-tooltip">{gitStore.gitLoading === 'push' ? 'Pushing...' : 'Push'}</span>
-              </button>
-            </div>
-          </div>
+<BottomBar {activeProfile} {appVersion} {claudePlan} {shellOpen} bind:profileMenuOpen bind:showSettings bind:settingsTab onToggleShell={toggleShell} onOpenExternal={openExternal} />
 
-          {#if gitStore.gitTab === 'changes'}
-            {#if gitStore.gitDiffFile}
-              <div class="git-diff-header">
-                <button class="git-diff-back" onclick={() => { gitStore.gitDiffFile = ''; gitStore.gitDiff = ''; }}>
-                  <svg width="12" height="12" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round"><polyline points="15 18 9 12 15 6"/></svg>
-                </button>
-                <span class="git-diff-filename">{gitStore.gitDiffFile}</span>
-              </div>
-              <div class="git-diff-view">
-                {#each gitStore.gitDiff.split('\n') as line}
-                  <div class="git-diff-line" class:diff-add={line.startsWith('+')} class:diff-del={line.startsWith('-')} class:diff-hunk={line.startsWith('@@')}>{line}</div>
-                {/each}
-              </div>
-            {:else if gitStore.gitFiles.length > 0}
-              <div class="git-file-list">
-                {#each gitStore.gitFiles as file}
-                  <!-- svelte-ignore a11y_no_static_element_interactions a11y_click_events_have_key_events -->
-                  <div class="git-file-item" onclick={() => viewDiff(file)}>
-                    <!-- svelte-ignore a11y_no_static_element_interactions a11y_click_events_have_key_events -->
-                    <span class="git-file-stage" onclick={(e) => { e.stopPropagation(); toggleStageFile(file); }} title={gitStore.stagedFiles.has(file.path) ? 'Unstage' : 'Stage'}>
-                      {#if gitStore.stagedFiles.has(file.path)}
-                        <svg width="12" height="12" viewBox="0 0 16 16" fill="var(--accent)"><path d="M13.78 4.22a.75.75 0 010 1.06l-7.25 7.25a.75.75 0 01-1.06 0L2.22 9.28a.75.75 0 011.06-1.06L6 10.94l6.72-6.72a.75.75 0 011.06 0z"/></svg>
-                      {:else}
-                        <svg width="12" height="12" viewBox="0 0 16 16" fill="none" stroke="var(--text-secondary)" stroke-width="1.5"><rect x="2" y="2" width="12" height="12" rx="2"/></svg>
-                      {/if}
-                    </span>
-                    <span class="git-file-status" class:modified={file.status === 'M'} class:added={file.status === 'A' || file.status === '??'} class:deleted={file.status === 'D'} class:renamed={file.status === 'R'}>{file.status === '??' ? 'U' : file.status}</span>
-                    <span class="git-file-path">{file.path}</span>
-                  </div>
-                {/each}
-              </div>
-              <div class="git-commit-row">
-                <input class="git-commit-input" type="text" bind:value={gitStore.gitCommitMsg} placeholder="Commit message..." onkeydown={(e) => { if (e.key === 'Enter' && gitStore.gitCommitMsg.trim()) doGitCommitStaged(); }} />
-                <button class="git-commit-btn" disabled={!gitStore.gitCommitMsg.trim() || gitStore.gitLoading === 'commit'} onclick={doGitCommitStaged}>
-                  {gitStore.gitLoading === 'commit' ? '...' : 'Commit'}
-                </button>
-              </div>
-            {:else}
-              <div class="git-empty">Working tree clean</div>
-            {/if}
-
-          {:else if gitStore.gitTab === 'history'}
-            <div class="git-history-list">
-              {#each gitStore.gitCommits as commit}
-                <div class="git-commit-item">
-                  <span class="git-commit-hash">{commit.short}</span>
-                  <span class="git-commit-message">{commit.message}</span>
-                  <span class="git-commit-date">{commit.date}</span>
-                </div>
-              {:else}
-                <div class="git-empty">No commits</div>
-              {/each}
-            </div>
-
-          {:else if gitStore.gitTab === 'branches'}
-            <div class="git-branch-list">
-              {#each gitStore.gitBranches as branch}
-                <!-- svelte-ignore a11y_no_static_element_interactions a11y_click_events_have_key_events -->
-                <div class="git-branch-item" class:current={branch.current} onclick={() => { if (!branch.current) switchBranch(branch.name); }}>
-                  {#if branch.current}
-                    <svg width="10" height="10" viewBox="0 0 16 16" fill="var(--accent)"><path d="M13.78 4.22a.75.75 0 010 1.06l-7.25 7.25a.75.75 0 01-1.06 0L2.22 9.28a.75.75 0 011.06-1.06L6 10.94l6.72-6.72a.75.75 0 011.06 0z"/></svg>
-                  {/if}
-                  <span class="git-branch-name-item">{branch.name}</span>
-                </div>
-              {:else}
-                <div class="git-empty">No branches</div>
-              {/each}
-            </div>
-          {/if}
-        </div>
-      {/if}
-    </div>
-    {/if}
-  </div>
-  <div class="bottom-center">
-    {#if usageStore.usageLimits}
-      {@const sColor = usageStore.usageLimits.sessionPercent > 80 ? '#f85149' : usageStore.usageLimits.sessionPercent > 50 ? '#d29922' : 'var(--accent)'}
-      {@const wColor = usageStore.usageLimits.weeklyAllPercent > 80 ? '#f85149' : usageStore.usageLimits.weeklyAllPercent > 50 ? '#d29922' : 'var(--accent)'}
-      <!-- svelte-ignore a11y_no_static_element_interactions a11y_click_events_have_key_events -->
-      <div class="usage-chips-clickable" onclick={() => { usageStore.showDashboard = true; usageStore.loadDashboard(); }}>
-        <div class="usage-chip"><span class="usage-dot" style="background:{sColor};box-shadow:0 0 6px {sColor}44;"></span><span class="usage-lbl">Session</span><span class="usage-val" style="color:{sColor}">{usageStore.usageLimits.sessionPercent.toFixed(0)}%</span></div>
-        <div class="usage-sep"></div>
-        <div class="usage-chip"><span class="usage-dot" style="background:{wColor};box-shadow:0 0 6px {wColor}44;"></span><span class="usage-lbl">Weekly</span><span class="usage-val" style="color:{wColor}">{usageStore.usageLimits.weeklyAllPercent.toFixed(0)}%</span></div>
-        {#if usageStore.usageLimits.weeklySonnetPercent != null}
-          {@const snColor = usageStore.usageLimits.weeklySonnetPercent > 80 ? '#f85149' : usageStore.usageLimits.weeklySonnetPercent > 50 ? '#d29922' : 'var(--accent)'}
-          <div class="usage-sep"></div>
-          <div class="usage-chip"><span class="usage-dot" style="background:{snColor};box-shadow:0 0 6px {snColor}44;"></span><span class="usage-lbl">Sonnet</span><span class="usage-val" style="color:{snColor}">{usageStore.usageLimits.weeklySonnetPercent.toFixed(0)}%</span></div>
-        {/if}
-      </div>
-    {:else}
-      <!-- svelte-ignore a11y_no_static_element_interactions a11y_click_events_have_key_events -->
-      <span class="limit-loading" onclick={() => { usageStore.showDashboard = true; usageStore.loadDashboard(); }} style="cursor:pointer;">
-        Set up usage tracking
-      </span>
-    {/if}
-  </div>
-  <div class="bottom-right">
-    <button class="shell-toggle-btn" class:active={shellOpen} disabled={!activeProfile && !shellOpen} onclick={toggleShell} title="Toggle shell (Cmd+L)">
-      <svg width="14" height="14" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round">
-        <polyline points="4 17 10 11 4 5"></polyline>
-        <line x1="12" y1="19" x2="20" y2="19"></line>
-      </svg>
-    </button>
-    {#if appVersion}<span class="bottom-version">v{appVersion}</span>{/if}
-  </div>
-</div>
-
-{#if updater.updateReady && !updater.updateDismissed}
-  <div class="update-notif">
-    <div class="un-header">
-      <svg class="un-icon" viewBox="0 0 24 24"><circle cx="12" cy="12" r="10"/><path d="M12 16v-4M12 8h.01"/></svg>
-      <div class="un-text">
-        <span class="un-title">Clauge v{updater.updateReady.version} is available</span>
-        <span class="un-desc">A new version has been downloaded. Restart to apply.</span>
-      </div>
-      <button class="un-close" onclick={() => updater.updateDismissed = true}>&times;</button>
-    </div>
-    <div class="un-actions">
-      <button class="un-btn primary" onclick={() => { updater.restartToUpdate(); }}>Restart to Update</button>
-      <button class="un-btn secondary" onclick={() => openExternal('https://clauge.ssh-i.in/changelog.html')}>
-        What's New
-        <svg viewBox="0 0 24 24"><path d="M18 13v6a2 2 0 01-2 2H5a2 2 0 01-2-2V8a2 2 0 012-2h6"/><polyline points="15 3 21 3 21 9"/><line x1="10" y1="14" x2="21" y2="3"/></svg>
-      </button>
-    </div>
-  </div>
-{/if}
+<UpdateToast {appVersion} onOpenExternal={openExternal} />
 </div>
 
 {#if showModal}
@@ -1352,480 +1147,11 @@ Anti-patterns to avoid:
 </div>
 {/if}
 
-{#if showSettings}
-<!-- svelte-ignore a11y_no_static_element_interactions a11y_click_events_have_key_events -->
-<div class="modal-backdrop">
-  <div class="stg-modal">
-    <div class="stg-header">
-      <span class="stg-title">Settings</span>
-      <button class="stg-close" onclick={() => showSettings = false}>&times;</button>
-    </div>
-    <div class="stg-layout">
-      <div class="stg-tabs">
-        <button class="stg-tab" class:active={settingsTab === 'settings'} onclick={() => settingsTab = 'settings'}>
-          <svg viewBox="0 0 24 24"><circle cx="12" cy="12" r="10" fill="none" stroke="currentColor" stroke-width="1.6"/><path d="M12 3v1m0 16v1m-9-9h1m16 0h1m-2.636-6.364l-.707.707M6.343 17.657l-.707.707m0-12.728l.707.707m11.314 11.314l.707.707" stroke="currentColor" stroke-width="1.6" stroke-linecap="round"/></svg>
-          Appearance
-        </button>
-        <button class="stg-tab" class:active={settingsTab === 'plugins'} onclick={() => { settingsTab = 'plugins'; pluginsStore.loadClaudePlugins(); }}>
-          <svg viewBox="0 0 24 24"><path d="M12 2L2 7l10 5 10-5-10-5z"/><path d="M2 17l10 5 10-5"/><path d="M2 12l10 5 10-5"/></svg>
-          Plugins
-        </button>
-        <button class="stg-tab" class:active={settingsTab === 'contexts'} onclick={() => { settingsTab = 'contexts'; contextsStore.loadContextSnippets(); }}>
-          <svg viewBox="0 0 24 24"><path d="M14 2H6a2 2 0 00-2 2v16a2 2 0 002 2h12a2 2 0 002-2V8z"/><polyline points="14 2 14 8 20 8"/><line x1="16" y1="13" x2="8" y2="13"/><line x1="16" y1="17" x2="8" y2="17"/><polyline points="10 9 9 9 8 9"/></svg>
-          Contexts
-        </button>
-        <button class="stg-tab" class:active={settingsTab === 'about'} onclick={() => settingsTab = 'about'}>
-          <svg viewBox="0 0 24 24"><circle cx="12" cy="12" r="10"/><path d="M12 16v-4M12 8h.01"/></svg>
-          About
-        </button>
-      </div>
-      <div class="stg-content">
+<SettingsModal bind:show={showSettings} bind:settingsTab {appVersion} onOpenExternal={openExternal} onApplyTheme={applyTheme} onApplyAccent={applyAccent} />
 
-    {#if settingsTab === 'settings'}
-      <div class="stg-section">
-        <div class="stg-section-label">Appearance</div>
-        <div class="stg-field">
-          <span class="stg-label">Theme</span>
-          <div class="chips">
-            <button class="chip" class:selected={theme.currentTheme === 'dark'} onclick={() => applyTheme('dark')}>Dark</button>
-            <button class="chip" class:selected={theme.currentTheme === 'light'} onclick={() => applyTheme('light')}>Light</button>
-          </div>
-        </div>
-        <div class="stg-field">
-          <span class="stg-label">Accent Color</span>
-          <div class="accent-row">
-            {#each ['#58a6ff', '#d2a8ff', '#3fb950', '#f85149', '#d29922', '#ff7b72'] as color}
-              <button class="color-dot" style="background:{color};{theme.accentColor === color ? 'box-shadow:0 0 0 2px var(--text-primary);' : ''}"
-                onclick={() => applyAccent(color)} title={color}></button>
-            {/each}
-          </div>
-        </div>
-      </div>
+<!-- UpdateToast handles both the toast and the what's new modal -->
 
-      <div class="stg-section">
-        <div class="stg-section-label">Terminal</div>
-        <div class="stg-field">
-          <span class="stg-label">Font Size</span>
-          <div style="display:flex;align-items:center;gap:8px;">
-            <input type="range" min="10" max="18" step="1" bind:value={terminalStore.termFontSize} class="stg-range"
-              oninput={(e) => {
-                const size = parseInt(e.target.value);
-                terminalStore.termFontSize = size;
-                localStorage.setItem('clauge-font-size', String(size));
-                for (const [, en] of terminalStore.terminalMap) { if (en.term) { en.term.options.fontSize = size; try { en.fitAddon.fit(); } catch(_) {} } }
-                for (const [, en] of shellStore.shellMap) { if (en.term) { en.term.options.fontSize = size; try { en.fitAddon.fit(); } catch(_) {} } }
-              }} />
-            <span style="font-size:11px;color:var(--text-secondary);width:24px;text-align:right;font-variant-numeric:tabular-nums;">{terminalStore.termFontSize}px</span>
-          </div>
-        </div>
-      </div>
-
-    {:else if settingsTab === 'plugins'}
-      <div class="plugin-subtabs">
-        <button class="plugin-subtab" class:active={pluginsStore.pluginTab === 'installed'} onclick={() => pluginsStore.pluginTab = 'installed'}>Installed ({pluginsStore.claudePlugins.length})</button>
-        <button class="plugin-subtab" class:active={pluginsStore.pluginTab === 'marketplace'} onclick={() => pluginsStore.pluginTab = 'marketplace'}>Marketplace</button>
-      </div>
-      {#if pluginsStore.pluginMsg}
-        <div class="plugin-msg" class:error={pluginsStore.pluginMsg.startsWith('Failed')}>{pluginsStore.pluginMsg}</div>
-      {/if}
-
-      {#if pluginsStore.pluginTab === 'installed'}
-        {#if pluginsStore.claudePlugins.length === 0}
-          <div class="plugin-empty">
-            <svg width="32" height="32" viewBox="0 0 24 24" fill="none" stroke="var(--border)" stroke-width="1.5" stroke-linecap="round" stroke-linejoin="round"><path d="M12 2L2 7l10 5 10-5-10-5z"/><path d="M2 17l10 5 10-5"/><path d="M2 12l10 5 10-5"/></svg>
-            <p>No plugins installed</p>
-            <button class="plugin-browse-btn" onclick={() => pluginsStore.pluginTab = 'marketplace'}>Browse Marketplace</button>
-          </div>
-        {:else}
-          <div class="plugins-list">
-            {#each pluginsStore.claudePlugins as plugin}
-              <div class="plugin-card">
-                <div class="plugin-icon">{plugin.name.charAt(0).toUpperCase()}</div>
-                <div class="plugin-info">
-                  <span class="plugin-name">{plugin.name}</span>
-                  <span class="plugin-cmd">{plugin.marketplace}{plugin.version && plugin.version !== 'unknown' ? ` · v${plugin.version}` : ''}</span>
-                </div>
-                <div class="plugin-actions">
-                  <button class="toggle-switch plugin-toggle" class:on={plugin.enabled} onclick={() => pluginsStore.togglePlugin(plugin)}>
-                    <span class="toggle-knob"></span>
-                  </button>
-                  <button class="plugin-uninstall" onclick={() => pluginsStore.pluginUninstallConfirm = plugin} title="Uninstall">
-                    <svg width="12" height="12" viewBox="0 0 16 16" fill="currentColor"><path d="M6.5 1.75a.25.25 0 01.25-.25h2.5a.25.25 0 01.25.25V3h-3V1.75zM11 3V1.75A1.75 1.75 0 009.25 0h-2.5A1.75 1.75 0 005 1.75V3H2.75a.75.75 0 000 1.5h.928l.747 10.218A1.75 1.75 0 006.172 16h3.656a1.75 1.75 0 001.747-1.282L12.322 4.5h.928a.75.75 0 000-1.5H11z"/></svg>
-                  </button>
-                </div>
-              </div>
-            {/each}
-          </div>
-        {/if}
-      {:else}
-        <div style="margin-bottom:12px;">
-          <input class="plugin-search full" type="text" bind:value={pluginsStore.pluginSearch} placeholder="Search plugins..." />
-        </div>
-        <div class="plugins-list marketplace">
-          {#each pluginsStore.marketplacePlugins.filter(p => !p.installed && (!pluginsStore.pluginSearch || p.name.toLowerCase().includes(pluginsStore.pluginSearch.toLowerCase()) || (p.description || '').toLowerCase().includes(pluginsStore.pluginSearch.toLowerCase()))) as plugin}
-            <div class="plugin-card">
-              <div class="plugin-icon mp">{plugin.name.charAt(0).toUpperCase()}</div>
-              <div class="plugin-info">
-                <span class="plugin-name">{plugin.name}</span>
-                <span class="plugin-cmd">{plugin.description || ''}</span>
-              </div>
-              {#if plugin.installs}
-                <span class="plugin-installs">{plugin.installs >= 1000 ? `${(plugin.installs / 1000).toFixed(0)}k` : plugin.installs}</span>
-              {/if}
-              <button class="plugin-install-btn" disabled={pluginsStore.installingPlugin === plugin.name} onclick={() => pluginsStore.installPlugin(plugin)}>
-                {pluginsStore.installingPlugin === plugin.name ? 'Installing...' : 'Install'}
-              </button>
-            </div>
-          {:else}
-            <div class="plugin-empty">
-              <p>No plugins found</p>
-            </div>
-          {/each}
-        </div>
-      {/if}
-
-    {:else if settingsTab === 'contexts'}
-      <div class="stg-section">
-        <div class="stg-section-label" style="display:flex;align-items:center;justify-content:space-between;">
-          Saved Contexts ({contextsStore.contextSnippets.length})
-          <button class="save-key-btn" style="font-size:10px;padding:3px 10px;" onclick={() => { contextsStore.contextEditing = { name: '', content: '' }; contextsStore.contextNewName = ''; contextsStore.contextNewContent = ''; }}>+ New</button>
-        </div>
-
-        {#if contextsStore.contextEditing}
-          <div class="ctx-editor">
-            <input type="text" class="ctx-name-input" bind:value={contextsStore.contextNewName} placeholder="Context name..." />
-            <textarea class="ctx-content-input" bind:value={contextsStore.contextNewContent} placeholder="Write your context, rules, or instructions..." rows="6"></textarea>
-            <div style="display:flex;gap:6px;justify-content:flex-end;">
-              <button class="save-key-btn" style="color:var(--text-secondary);border-color:var(--border);" onclick={() => contextsStore.contextEditing = null}>Cancel</button>
-              <button class="save-key-btn" disabled={!contextsStore.contextNewName.trim() || !contextsStore.contextNewContent.trim()} onclick={() => contextsStore.saveContextSnippet()}>Save</button>
-            </div>
-          </div>
-        {/if}
-
-        <div class="ctx-list">
-          {#each contextsStore.contextSnippets as ctx}
-            {#if !contextsStore.contextEditing || contextsStore.contextEditing.name !== ctx.name}
-            <div class="ctx-card">
-              <div class="ctx-card-info">
-                <span class="ctx-card-name">{ctx.name}</span>
-                <span class="ctx-card-preview">{ctx.preview}</span>
-              </div>
-              <div class="ctx-card-actions">
-                <button class="ctx-action-btn" onclick={() => { contextsStore.contextEditing = ctx; contextsStore.contextNewName = ctx.name; contextsStore.contextNewContent = ctx.content; }} title="Edit">
-                  <svg width="11" height="11" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round"><path d="M11 4H4a2 2 0 00-2 2v14a2 2 0 002 2h14a2 2 0 002-2v-7"/><path d="M18.5 2.5a2.121 2.121 0 013 3L12 15l-4 1 1-4 9.5-9.5z"/></svg>
-                </button>
-                <button class="ctx-action-btn danger" onclick={() => contextsStore.deleteContextSnippet(ctx.name)} title="Delete">
-                  <svg width="11" height="11" viewBox="0 0 16 16" fill="currentColor"><path d="M6.5 1.75a.25.25 0 01.25-.25h2.5a.25.25 0 01.25.25V3h-3V1.75zM11 3V1.75A1.75 1.75 0 009.25 0h-2.5A1.75 1.75 0 005 1.75V3H2.75a.75.75 0 000 1.5h.928l.747 10.218A1.75 1.75 0 006.172 16h3.656a1.75 1.75 0 001.747-1.282L12.322 4.5h.928a.75.75 0 000-1.5H11z"/></svg>
-                </button>
-              </div>
-            </div>
-            {/if}
-          {:else}
-            {#if !contextsStore.contextEditing}
-              <div style="padding:20px;text-align:center;font-size:12px;color:var(--text-secondary);">
-                No contexts yet. Create one to attach to sessions.
-              </div>
-            {/if}
-          {/each}
-        </div>
-      </div>
-
-    {:else if settingsTab === 'about'}
-      <div class="about-content">
-        <div class="about-header">
-          <span class="about-app-name">Clauge</span>
-          <span class="about-version">v{appVersion || '1.0.0'}</span>
-        </div>
-        <p class="about-desc">A developer toolkit for managing sessions, terminals, and workflows — all in one window.</p>
-
-        <div class="about-section-label">TECH STACK</div>
-        <div class="about-tech-grid">
-          <span class="about-tech-pill">
-            <svg viewBox="0 0 24 24" class="tech-icon"><circle cx="12" cy="12" r="3"/><path d="M12.22 2h-.44a2 2 0 00-2 2v.18a2 2 0 01-1 1.73l-.43.25a2 2 0 01-2 0l-.15-.08a2 2 0 00-2.73.73l-.22.38a2 2 0 00.73 2.73l.15.1a2 2 0 011 1.72v.51a2 2 0 01-1 1.74l-.15.09a2 2 0 00-.73 2.73l.22.38a2 2 0 002.73.73l.15-.08a2 2 0 012 0l.43.25a2 2 0 011 1.73V20a2 2 0 002 2h.44a2 2 0 002-2v-.18a2 2 0 011-1.73l.43-.25a2 2 0 012 0l.15.08a2 2 0 002.73-.73l.22-.39a2 2 0 00-.73-2.73l-.15-.08a2 2 0 01-1-1.74v-.5a2 2 0 011-1.74l.15-.09a2 2 0 00.73-2.73l-.22-.38a2 2 0 00-2.73-.73l-.15.08a2 2 0 01-2 0l-.43-.25a2 2 0 01-1-1.73V4a2 2 0 00-2-2z"/></svg>
-            Rust
-          </span>
-          <span class="about-tech-pill">
-            <svg viewBox="0 0 24 24" class="tech-icon"><circle cx="12" cy="12" r="10"/><circle cx="12" cy="12" r="4"/></svg>
-            Tauri v2
-          </span>
-          <span class="about-tech-pill">
-            <svg viewBox="0 0 24 24" class="tech-icon"><path d="M12.1 2L1 21h22L12.1 2z" fill="none" stroke="currentColor" stroke-width="1.5" stroke-linejoin="round"/></svg>
-            SvelteKit
-          </span>
-          <span class="about-tech-pill">
-            <svg viewBox="0 0 24 24" class="tech-icon"><polyline points="4 17 10 11 4 5"/><line x1="12" y1="19" x2="20" y2="19"/></svg>
-            xterm.js
-          </span>
-        </div>
-
-        <div class="about-section-label">LINKS</div>
-        <div class="about-links">
-          <!-- svelte-ignore a11y_no_static_element_interactions a11y_click_events_have_key_events -->
-          <span class="about-link-btn" onclick={() => openExternal('https://github.com/ansxuman/Clauge')}>
-            <svg viewBox="0 0 24 24"><path d="M9 19c-5 1.5-5-2.5-7-3m14 6v-3.87a3.37 3.37 0 00-.94-2.61c3.14-.35 6.44-1.54 6.44-7A5.44 5.44 0 0020 4.77 5.07 5.07 0 0019.91 1S18.73.65 16 2.48a13.38 13.38 0 00-7 0C6.27.65 5.09 1 5.09 1A5.07 5.07 0 005 4.77a5.44 5.44 0 00-1.5 3.78c0 5.42 3.3 6.61 6.44 7A3.37 3.37 0 009 18.13V22"/></svg>
-            <span>GitHub</span>
-          </span>
-          <!-- svelte-ignore a11y_no_static_element_interactions a11y_click_events_have_key_events -->
-          <span class="about-link-btn" onclick={() => openExternal('https://github.com/ansxuman/Clauge/issues')}>
-            <svg viewBox="0 0 24 24"><circle cx="12" cy="12" r="10"/><path d="M12 8v4M12 16h.01"/></svg>
-            <span>Report Issue</span>
-          </span>
-          <!-- svelte-ignore a11y_no_static_element_interactions a11y_click_events_have_key_events -->
-          <span class="about-link-btn" onclick={() => openExternal('https://github.com/ansxuman')}>
-            <svg viewBox="0 0 24 24"><path d="M20 21v-2a4 4 0 00-4-4H8a4 4 0 00-4 4v2"/><circle cx="12" cy="7" r="4"/></svg>
-            <span>Developer</span>
-          </span>
-          <!-- svelte-ignore a11y_no_static_element_interactions a11y_click_events_have_key_events -->
-          <span class="about-link-btn" onclick={() => openExternal('https://clauge.ssh-i.in')}>
-            <svg viewBox="0 0 24 24"><circle cx="12" cy="12" r="10"/><line x1="2" y1="12" x2="22" y2="12"/><path d="M12 2a15.3 15.3 0 014 10 15.3 15.3 0 01-4 10 15.3 15.3 0 01-4-10 15.3 15.3 0 014-10z"/></svg>
-            <span>Website</span>
-          </span>
-        </div>
-
-        <div class="about-section-label">SUPPORT</div>
-        <!-- svelte-ignore a11y_no_static_element_interactions a11y_click_events_have_key_events -->
-        <span class="about-coffee" onclick={() => openExternal('https://buymeacoffee.com/ansxuman')}>
-          <svg viewBox="0 0 24 24"><path d="M17 8h1a4 4 0 110 8h-1"/><path d="M3 8h14v9a4 4 0 01-4 4H7a4 4 0 01-4-4V8z"/><line x1="6" y1="2" x2="6" y2="4"/><line x1="10" y1="2" x2="10" y2="4"/><line x1="14" y1="2" x2="14" y2="4"/></svg>
-          Buy me a coffee
-        </span>
-
-      </div>
-    {/if}
-      </div>
-    </div>
-  </div>
-</div>
-{/if}
-
-{#if updater.showWhatsNew}
-<div class="modal-backdrop">
-  <div class="modal whats-new-modal">
-    {#if updater.updateReady}
-      <h2>v{updater.updateReady.version}</h2>
-      <div class="whats-new-body">{@html (updater.updateReady.body || '')
-        .replace(/\r\n/g, '\n')
-        .replace(/^### (.+)$/gm, '<h4>$1</h4>')
-        .replace(/^## (.+)$/gm, '<h3>$1</h3>')
-        .replace(/\*\*([^*]+)\*\*/g, '<strong>$1</strong>')
-        .replace(/`([^`]+)`/g, '<code>$1</code>')
-        .replace(/^\s*[-*] (.+)$/gm, '<li>$1</li>')
-        .replace(/((?:<li>.*<\/li>\n?)+)/g, '<ul>$1</ul>')
-        .replace(/\n\n+/g, '<br>')
-        .replace(/\n/g, '<br>')
-      }</div>
-      <div class="modal-actions">
-        <button onclick={() => updater.showWhatsNew = false}>Later</button>
-        <button class="create-btn" onclick={() => { updater.showWhatsNew = false; updater.restartToUpdate(); }}>Restart</button>
-      </div>
-    {:else}
-      <h2>What's New in v{appVersion}</h2>
-      <div class="whats-new-body">{@html updater.whatsNewBody
-        .replace(/\r\n/g, '\n')
-        .replace(/^### (.+)$/gm, '<h4>$1</h4>')
-        .replace(/^## (.+)$/gm, '<h3>$1</h3>')
-        .replace(/\*\*([^*]+)\*\*/g, '<strong>$1</strong>')
-        .replace(/`([^`]+)`/g, '<code>$1</code>')
-        .replace(/^\s*[-*] (.+)$/gm, '<li>$1</li>')
-        .replace(/((?:<li>.*<\/li>\n?)+)/g, '<ul>$1</ul>')
-        .replace(/\n\n+/g, '<br>')
-        .replace(/\n/g, '<br>')
-      }</div>
-      <div class="modal-actions">
-        <button onclick={() => updater.showWhatsNew = false}>Got it</button>
-      </div>
-    {/if}
-  </div>
-</div>
-{/if}
-
-{#if usageStore.showDashboard}
-<!-- svelte-ignore a11y_no_static_element_interactions a11y_click_events_have_key_events -->
-<div class="modal-backdrop">
-  <div class="dash-modal">
-    <div class="dash-header">
-      <span class="dash-title">Usage Dashboard {#if claudePlan}<span class="dash-plan-badge">{claudePlan}</span>{/if}</span>
-      <div class="dash-header-right">
-        <select class="dash-period" bind:value={usageStore.dashboardDays} onchange={() => usageStore.loadDashboard()}>
-          <option value={7}>7 days</option>
-          <option value={30}>30 days</option>
-          <option value={90}>90 days</option>
-          <option value={9999}>All time</option>
-        </select>
-        <button class="stg-close" onclick={() => usageStore.showDashboard = false}>&times;</button>
-      </div>
-    </div>
-    {#if usageStore.dashboardLoading}
-      <div class="dash-loading"><div class="dash-spinner"></div>Analyzing sessions...</div>
-    {:else if usageStore.dashboardData}
-      <div class="dash-body">
-        <div class="dash-stats">
-          <div class="dash-stat"><span class="dash-stat-value">{usageStore.formatCost(usageStore.dashboardData.totalCost)}</span><span class="dash-stat-label">Total Cost</span></div>
-          <div class="dash-stat"><span class="dash-stat-value">{usageStore.dashboardData.totalApiCalls.toLocaleString()}</span><span class="dash-stat-label">API Calls</span></div>
-          <div class="dash-stat"><span class="dash-stat-value">{usageStore.dashboardData.cacheHitPercent.toFixed(1)}%</span><span class="dash-stat-label">Cache Hit</span></div>
-          <div class="dash-stat"><span class="dash-stat-value">{usageStore.dashboardData.totalSessions}</span><span class="dash-stat-label">Sessions</span></div>
-        </div>
-
-        <!-- Row 2: Tokens -->
-        <div class="dash-tokens-bar">
-          <span><strong>In:</strong> {usageStore.formatTokens(usageStore.dashboardData.totalInputTokens)}</span>
-          <span><strong>Out:</strong> {usageStore.formatTokens(usageStore.dashboardData.totalOutputTokens)}</span>
-          <span><strong>Cache R:</strong> {usageStore.formatTokens(usageStore.dashboardData.totalCacheReadTokens)}</span>
-          <span><strong>Cache W:</strong> {usageStore.formatTokens(usageStore.dashboardData.totalCacheWriteTokens)}</span>
-        </div>
-
-        <!-- Row 3: Chart -->
-        {#if usageStore.dashboardData.daily.length > 0}
-          <div class="dash-section">
-            <div class="dash-section-label">Daily Activity</div>
-            <div class="dash-chart">
-              {#each usageStore.dashboardData.daily.slice(-21) as day}
-                {@const mc = Math.max(...usageStore.dashboardData.daily.slice(-21).map(d => d.cost), 0.01)}
-                <div class="dash-bar-wrap" title="{day.date}: {usageStore.formatCost(day.cost)} · {day.calls} calls">
-                  <div class="dash-bar" style="height:{Math.max(3, (day.cost / mc) * 100)}%"></div>
-                  <span class="dash-bar-label">{day.date.slice(8)}</span>
-                </div>
-              {/each}
-            </div>
-          </div>
-        {/if}
-
-        <!-- Row 4: Live Usage + Models -->
-        <div class="dash-grid">
-          <div class="dash-section">
-            <div class="dash-section-label">Live Usage</div>
-            {#if usageStore.usageLimits}
-              <div style="display:flex;flex-direction:column;gap:8px;">
-                <div class="dash-live-row">
-                  <span class="dash-live-lbl">Session</span>
-                  <div class="dash-live-bar"><div style="width:{usageStore.usageLimits.sessionPercent}%;background:{usageStore.usageLimits.sessionPercent > 80 ? '#f85149' : usageStore.usageLimits.sessionPercent > 50 ? '#d29922' : 'var(--accent)'};height:100%;border-radius:2px;"></div></div>
-                  <span class="dash-live-pct" style="color:{usageStore.usageLimits.sessionPercent > 80 ? '#f85149' : usageStore.usageLimits.sessionPercent > 50 ? '#d29922' : 'var(--accent)'}">{usageStore.usageLimits.sessionPercent.toFixed(1)}%</span>
-                </div>
-                <div class="dash-live-row">
-                  <span class="dash-live-lbl">Weekly</span>
-                  <div class="dash-live-bar"><div style="width:{usageStore.usageLimits.weeklyAllPercent}%;background:{usageStore.usageLimits.weeklyAllPercent > 80 ? '#f85149' : usageStore.usageLimits.weeklyAllPercent > 50 ? '#d29922' : 'var(--accent)'};height:100%;border-radius:2px;"></div></div>
-                  <span class="dash-live-pct" style="color:{usageStore.usageLimits.weeklyAllPercent > 80 ? '#f85149' : usageStore.usageLimits.weeklyAllPercent > 50 ? '#d29922' : 'var(--accent)'}">{usageStore.usageLimits.weeklyAllPercent.toFixed(1)}%</span>
-                </div>
-                {#if usageStore.usageLimits.weeklySonnetPercent != null}
-                  <div class="dash-live-row">
-                    <span class="dash-live-lbl">Sonnet</span>
-                    <div class="dash-live-bar"><div style="width:{usageStore.usageLimits.weeklySonnetPercent}%;background:{usageStore.usageLimits.weeklySonnetPercent > 80 ? '#f85149' : usageStore.usageLimits.weeklySonnetPercent > 50 ? '#d29922' : 'var(--accent)'};height:100%;border-radius:2px;"></div></div>
-                    <span class="dash-live-pct" style="color:{usageStore.usageLimits.weeklySonnetPercent > 80 ? '#f85149' : usageStore.usageLimits.weeklySonnetPercent > 50 ? '#d29922' : 'var(--accent)'}">{usageStore.usageLimits.weeklySonnetPercent.toFixed(1)}%</span>
-                  </div>
-                {/if}
-              </div>
-              {#if usageStore.showKeyEdit}
-                <div style="margin-top:8px;">
-                  <input type="password" bind:value={usageStore.sessionKeyInput} placeholder="sk-ant-sid01-..." style="padding:5px 8px;border-radius:4px;border:1px solid var(--border);background:transparent;color:var(--text-primary);font-size:11px;width:100%;margin-bottom:4px;" />
-                  <div style="display:flex;gap:6px;">
-                    <button class="save-key-btn" onclick={async () => {
-                      if (usageStore.sessionKeyInput.trim()) {
-                        await invoke("save_session_key", { key: usageStore.sessionKeyInput.trim() });
-                        usageStore.sessionKeyConfigured = true; usageStore.showKeyEdit = false;
-                        await usageStore.loadUsageLimits();
-                      }
-                    }}>Save</button>
-                    <button class="save-key-btn" style="color:var(--text-secondary);border-color:var(--border);" onclick={() => usageStore.showKeyEdit = false}>Cancel</button>
-                  </div>
-                </div>
-              {:else}
-                <div style="display:flex;align-items:center;gap:6px;margin-top:8px;">
-                  <span style="font-size:9px;color:var(--text-secondary);">Refresh every</span>
-                  <select class="dash-refresh-select" bind:value={usageStore.usageRefreshMins} onchange={() => {
-                    localStorage.setItem('clauge-usage-refresh', String(usageStore.usageRefreshMins));
-                    usageStore.startRefreshInterval();
-                  }}>
-                    <option value={5}>5 min</option><option value={15}>15 min</option><option value={30}>30 min</option>
-                    <option value={60}>1 hour</option><option value={360}>6 hours</option><option value={720}>12 hours</option>
-                  </select>
-                  <button class="dash-edit-key" onclick={() => usageStore.showKeyEdit = true} title="Update session key">
-                    Session Key
-                    <svg width="10" height="10" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round"><path d="M11 4H4a2 2 0 00-2 2v14a2 2 0 002 2h14a2 2 0 002-2v-7"/><path d="M18.5 2.5a2.121 2.121 0 013 3L12 15l-4 1 1-4 9.5-9.5z"/></svg>
-                  </button>
-                </div>
-              {/if}
-            {:else}
-              <div style="padding:8px 0;">
-                <p style="font-size:11px;color:var(--text-secondary);margin:0 0 8px;">Connect to see live session limits</p>
-                <input type="password" bind:value={usageStore.sessionKeyInput} placeholder="sk-ant-sid01-..." style="padding:5px 8px;border-radius:4px;border:1px solid var(--border);background:transparent;color:var(--text-primary);font-size:11px;width:100%;margin-bottom:6px;" />
-                <p style="font-size:9px;color:var(--text-secondary);margin:0 0 6px;">claude.ai → DevTools → Cookies → sessionKey</p>
-                <button class="save-key-btn" onclick={async () => {
-                  if (usageStore.sessionKeyInput.trim()) {
-                    await invoke("save_session_key", { key: usageStore.sessionKeyInput.trim() });
-                    usageStore.sessionKeyConfigured = true; usageStore.usageError = '';
-                    await usageStore.loadUsageLimits();
-                    usageStore.startRefreshInterval();
-                  }
-                }}>Connect</button>
-              </div>
-            {/if}
-          </div>
-          <div class="dash-section">
-            <div class="dash-section-label">Models</div>
-            {#each usageStore.dashboardData.byModel as m}
-              <div class="dash-model-row">
-                <div class="dash-model-info"><span class="dash-model-name">{m.model}</span><span class="dash-model-meta">{m.calls} calls · {m.cacheHitPercent.toFixed(0)}% cache</span></div>
-                <span class="dash-model-cost">{usageStore.formatCost(m.cost)}</span>
-              </div>
-            {/each}
-          </div>
-        </div>
-
-        <!-- Row 5: Projects + Top Sessions -->
-        <div class="dash-grid">
-          <div class="dash-section">
-            <div class="dash-section-label">Projects ({usageStore.dashboardData.byProject.length})</div>
-            <div class="dash-scroll">
-              {#each usageStore.dashboardData.byProject as p}
-                <div class="dash-model-row">
-                  <div class="dash-model-info"><span class="dash-model-name" title={p.project}>{usageStore.decodeProjectName(p.project)}</span><span class="dash-model-meta">{p.sessions} sess · {p.calls} calls</span></div>
-                  <span class="dash-model-cost">{usageStore.formatCost(p.cost)}</span>
-                </div>
-              {/each}
-            </div>
-          </div>
-          <div class="dash-section">
-            <div class="dash-section-label">Top Sessions</div>
-            <div class="dash-scroll">
-              {#each usageStore.dashboardData.topSessions as s}
-                <div class="dash-model-row">
-                  <div class="dash-model-info"><span class="dash-model-name" title={s.project}>{usageStore.decodeProjectName(s.project)}</span><span class="dash-model-meta">{s.model} · {s.sessionId.slice(0, 8)}</span></div>
-                  <span class="dash-model-cost">{usageStore.formatCost(s.cost)}</span>
-                </div>
-              {/each}
-            </div>
-          </div>
-        </div>
-
-        <!-- Row 5: Tools + Shell -->
-        <div class="dash-grid">
-          <div class="dash-section">
-            <div class="dash-section-label">Tools</div>
-            {#each usageStore.dashboardData.tools.slice(0, 6) as t}
-              <div class="dash-tool-row">
-                <span class="dash-tool-name">{t.name}</span>
-                <div class="dash-tool-bar-bg"><div class="dash-tool-bar-fill" style="width:{Math.max(3, (t.count / (usageStore.dashboardData.tools[0]?.count || 1)) * 100)}%"></div></div>
-                <span class="dash-tool-count">{t.count.toLocaleString()}</span>
-              </div>
-            {/each}
-          </div>
-          <div class="dash-section">
-            <div class="dash-section-label">Shell</div>
-            {#each usageStore.dashboardData.shellCommands.slice(0, 6) as s}
-              <div class="dash-tool-row">
-                <span class="dash-tool-name" style="font-family:monospace;">{s.name}</span>
-                <div class="dash-tool-bar-bg"><div class="dash-tool-bar-fill" style="width:{Math.max(3, (s.count / (usageStore.dashboardData.shellCommands[0]?.count || 1)) * 100)}%"></div></div>
-                <span class="dash-tool-count">{s.count.toLocaleString()}</span>
-              </div>
-            {/each}
-          </div>
-        </div>
-      </div>
-    {:else}
-      <div class="dash-loading">No usage data found</div>
-    {/if}
-  </div>
-</div>
-{/if}
+<UsageDashboard {claudePlan} />
 
 {#if contextsStore.showContextPicker}
 <!-- svelte-ignore a11y_no_static_element_interactions a11y_click_events_have_key_events -->
@@ -1875,23 +1201,7 @@ Anti-patterns to avoid:
 </div>
 {/if}
 
-{#if deleteConfirm}
-<div class="modal-backdrop" style="animation:fadeIn 0.1s ease-out;">
-  <div class="modal" style="max-width:360px;animation:slideIn 0.15s ease-out;">
-    <div style="text-align:center;padding:20px 20px 0;">
-      <svg width="32" height="32" viewBox="0 0 16 16" fill="#f85149" style="margin-bottom:12px;"><path d="M6.5 1.75a.25.25 0 01.25-.25h2.5a.25.25 0 01.25.25V3h-3V1.75zM11 3V1.75A1.75 1.75 0 009.25 0h-2.5A1.75 1.75 0 005 1.75V3H2.75a.75.75 0 000 1.5h.928l.747 10.218A1.75 1.75 0 006.172 16h3.656a1.75 1.75 0 001.747-1.282L12.322 4.5h.928a.75.75 0 000-1.5H11zm-5.522 1.5l.735 10.06a.25.25 0 00.249.19h3.076a.25.25 0 00.249-.19l.735-10.06H5.478z"/></svg>
-      <h2 style="font-size:15px;margin-bottom:8px;">Delete this session?</h2>
-      <p style="font-size:13px;color:var(--text-secondary);line-height:1.5;">
-        Are you sure you want to delete the <strong style="color:var(--text-primary);">{deleteConfirm.projectName} — {deleteConfirm.purpose}</strong> session?
-      </p>
-    </div>
-    <div class="modal-actions" style="padding:16px 20px;">
-      <button onclick={() => deleteConfirm = null}>Cancel</button>
-      <button style="background:#f85149 !important;border-color:transparent !important;color:#fff !important;" onclick={confirmDelete}>Delete</button>
-    </div>
-  </div>
-</div>
-{/if}
+<DeleteConfirmModal profile={deleteConfirm} onConfirm={confirmDelete} onCancel={() => deleteConfirm = null} />
 
 {#if pluginsStore.pluginUninstallConfirm}
 <div class="modal-backdrop" style="z-index:1100;">
